requirements.txt
absl-py==0.14.0
addict==2.4.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.6
cityscapesScripts==2.2.0
colorama==0.4.4
coloredlogs==15.0.1
correlation-cuda==0.0.0
cycler==0.10.0
Cython==0.29.23
decorator==5.0.7
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
einops==0.4.1
entrypoints==0.3
et-xmlfile==1.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.4.0
humanfriendly==10.0
idna==3.2
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
llvmlite==0.37.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
mmcv-full==1.4.1
mmsegmentation==0.20.2
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numba==0.54.0
numpy @ file:///D:/bld/numpy_1616933441329/work
oauthlib==3.1.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.3.56
openpyxl==3.0.10
packaging==20.9
pandas @ file:///C:/ci/pandas_1618365634936/work
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1617386341487/work
pluggy==1.0.0
POT==0.7.0
pretrainedmodels==0.7.4
prettytable==2.5.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydensecrf @ file:///C:/Users/wp/pydensecrf-1.0rc2-cp38-cp38-win_amd64.whl
Pygments==2.8.1
pyparsing==2.4.7
pyquaternion==0.9.9
pyreadline3==3.3
pyrsistent==0.17.3
pytest==6.2.5
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.3
seaborn==0.11.2
segmentation-models-pytorch==0.3.1
Send2Trash==1.5.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
spatial-correlation-sampler==0.3.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX @ file:///tmp/build/80754af9/tensorboardx_1621440489103/work
terminado==0.9.4
testpath==0.4.4
thop==0.1.1.post2209072238
threadpoolctl==2.1.0
timm==0.4.12
toml==0.10.2
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchmetrics==0.11.0
torchsummary==1.5.1
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
tslearn==0.5.2
typing==3.7.4.3
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wincertstore==0.2
yapf==0.31.0
